[tests/t*random*.c]   * Replaced some fprintf with stderr by printf (but not those     concerning memory allocation errors).   * Do bitwise operations in unsigned arithmetic.   * Formatting.
[From SVN r12189 (trunk)]

+++ tests/tgrandom.c
@@ -32,19 +32,19 @@ test_special (mpfr_prec_t p)
   mpfr_init2 (x, p);
 
   inexact = mpfr_grandom (x, NULL, RANDS, MPFR_RNDN);
-  if ((inexact & 3) == 0)
+  if (((unsigned int) inexact & 3) == 0)
     {
       printf ("Error: mpfr_grandom() returns a zero ternary value.\n");
       exit (1);
     }
-  if ((inexact & (3 << 2)) != 0)
+  if (((unsigned int) inexact & (3 << 2)) != 0)
     {
       printf ("Error: the second ternary value of mpfr_grandom(x, NULL, ...)"
               " must be 0.\n");
       exit (1);
     }
 
-  mpfr_clear(x);
+  mpfr_clear (x);
 }
 
 
@@ -64,7 +64,8 @@ test_grandom (long nbtests, mpfr_prec_t prec, mpfr_rnd_t rnd,
   for (i = 0; i < nbtests; i += 2)
     {
       inexact = mpfr_grandom (t[i], t[i + 1], RANDS, MPFR_RNDN);
-      if ((inexact & 3) == 0 || (inexact & (3 << 2)) == 0)
+      if (((unsigned int) inexact & 3) == 0 ||
+          ((unsigned int) inexact & (3 << 2)) == 0)
         {
           /* one call in the loop pretended to return an exact number! */
           printf ("Error: mpfr_grandom() returns a zero ternary value.\n");
@@ -113,6 +114,7 @@ main (int argc, char *argv[])
 {
   long nbtests;
   int verbose;
+
   tests_start_mpfr ();
 
   verbose = 0;
